dgb-py: solved overflow
Total number of samples may exceed MAX(int32). Switching to int64.

+++ deeplearning_apply_clientlib.py
@@ -94,7 +94,7 @@ class Message:
         ret = list()
         offset = 0
         for shape,dtype in zip(shapes,dtypes):
-          nrsamples = np.prod(shape)
+          nrsamples = np.prod(shape,dtype=np.int64)
           arr = np.frombuffer(arrptr,dtype,count=nrsamples,offset=offset)
           arr = arr.reshape( shape )
           offset += arr.nbytes
